<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> bbe4d9cf03dcd14dc250f9c2de17d39c14fb1c1b
% Please edit documentation in R/getCircomics_doc.R
\name{attriColorGene}
\alias{attriColorGene}
\title{Attribute Color to Gene}
\usage{
attriColorGene(df)
}
\arguments{
\item{df}{data frame with mRNA or CNA or mutation frequency or methylation (numeric).}
}
\value{
A list colors for every gene
}
\description{
Attribute Color to Gene
}
\examples{
\dontrun{`
load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
colorsls <- attriColorGene(ListProfData_bkp$Expression$luad_tcga_pub)
}
}

