% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCircomics_doc.R
\name{attriColorValue}
\alias{attriColorValue}
\title{Attribute Color to Value}
\usage{
attriColorValue(Value, df, colors = c(a, b, c, d, e), feet)
}
\arguments{
\item{Value}{integer}

\item{df}{data frame with numeric values}

\item{colors}{a vector of 5 colors}

\item{feet}{the interval between two successive colors in the palette (0.1)}
}
\value{
Hex Color Code
}
\description{
Attribute Color to Value
}
\examples{
\dontrun{
load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
colorRef <- attriColorValue(500, ListProfData_bkp$Expression$luad_tcga_pub, colors = c("blue", "white","yellow","red", "black"), feet=10)
}
}

