% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome.R
\name{attriColorVector}
\alias{attriColorVector}
\title{Attribute color to a vector of numeric values}
\usage{
attriColorVector(Value, vector, colors=c(a,b,c),feet)
}
\arguments{
\item{Value}{numeric}

\item{vector}{A vector of numeric data}

\item{colors}{3 colors}

\item{feet}{An interval between two numeric value needed to change the color}
}
\value{
A vetor of colors
}
\description{
Attribute color to a vector of numeric values
}
\examples{
cgds <- cBioPortal(
hostname = "www.cbioportal.org",
protocol = "https",
api = "/api/v2/api-docs"
)
\dontrun{
getDataByGenes( api =  cgds,
studyId = "gbm_tcga_pub",
genes = c("NF1", "TP53", "ABL1"),
by = "hugoGeneSymbol",
molecularProfileIds = "gbm_tcga_pub_mrna"
)
}
}
