% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_doc.R
\name{attriShape2Node}
\alias{attriShape2Node}
\title{Attributes shape to Nodes}
\usage{
attriShape2Node(gene, genelist)
}
\arguments{
\item{gene}{symbol "TP53"}

\item{genelist}{a vector of gene symbol}
}
\value{
A data frame with egdes attributes
}
\description{
Attributes shape to Nodes
}
\examples{

genelist <- c("DKK3" , "NBN"  , "MYO6" , "TP53" , "PML"  , "IFI16" ,"BRCA1")
NodeShape <- attriShape2Gene("DKK3", genelist)

}

