% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffeewheel.R
\name{coffeewheel}
\alias{coffeewheel}
\title{This is an htmlwidgets-based visualization tool for hierarchical data.
It is zoomable, meaning that you can interact with the hierarchy and zoom in/out accordingly.}
\usage{
coffeewheel(treeData, width=600, height=600, main="", partitionAttribute="value")
}
\arguments{
\item{treeData}{A hierarchical tree data as in example}

\item{width}{600}

\item{height}{600}

\item{main}{Title}

\item{partitionAttribute}{"value"}
}
\value{
A circular layout with genetic profile.
}
\description{
This is an htmlwidgets-based visualization tool for hierarchical data.
It is zoomable, meaning that you can interact with the hierarchy and zoom in/out accordingly.
}
\examples{
load(paste(path.package("bioCancer"),"/extdata/ListProfData.RData", sep=""))
bioCancer::bioCancer()

}

