% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffeewheel.R
\name{coffeewheelOutput}
\alias{coffeewheelOutput}
\title{Widget output function for use in Shiny}
\usage{
coffeewheelOutput(outputId, width = 1000, height = 1000)
}
\description{
Widget output function for use in Shiny
}

