% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_doc.R
\name{diagrammeR}
\alias{diagrammeR}
\title{Plot network with nodes and edges attributes}
\usage{
diagrammeR(NodeAttri_Reactome,NodeAttri_Classifier,NodeAttri_ProfData)
}
\arguments{
\item{NodeAttri_Reactome}{Node attribute from Reactome database ('Freq. Interaction')}

\item{NodeAttri_Classifier}{Node attributes from geNetClassifier ('mRNA','Studies','mRNA/Studies')}

\item{NodeAttri_ProfData}{Node attributes from Profiles Data ('Mutation', 'CNA', 'Met_HM450', 'Met_HM27')}
}
\value{
plot
}
\description{
Plot network with nodes and edges attributes
}
\examples{
example <- "runManually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
##load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
ReactomeFI <- readRDS(paste0(system.file(package = "bioCancer"),"/extdata/DisGeNet.RDS"))
diagrammeR('Freq.Interaction', 'mRNA', 'Met_HM450')
}
}

