% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portal.R
\name{displayTable}
\alias{displayTable}
\title{Display dataframe in table using DT package}
\usage{
displayTable(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
A table
}
\description{
Display dataframe in table using DT package
}
\examples{
cgds <- cBioPortal(
hostname = "www.cbioportal.org",
protocol = "https",
api = "/api/v2/api-docs"
)
\dontrun{
getDataByGenes( api =  cgds,
studyId = "gbm_tcga_pub",
genes = c("NF1", "TP53", "ABL1"),
by = "hugoGeneSymbol",
molecularProfileIds = "gbm_tcga_pub_mrna"
)
}
}
