% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationFuncs.R
\name{.getTableName}
\alias{.getTableName}
\title{Gets the table name from the INPARANOID style genus names.}
\usage{
.getTableName(genus)
}
\arguments{
\item{genus}{5 character INPARANOID genus name, such as "BOSTA", "HOMSA" or "MUSMU".}
}
\value{
Table name for genus.
}
\description{
Gets the table name from the INPARANOID style genus names.
}
\references{
\url{http://www.bioconductor.org/packages/release/bioc/html/AnnotationDbi.html}
}
\author{
Stefan McKinnon Edwards \email{stefanm.edwards@agrsci.dk}
}
