% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationFuncs.R
\name{.pickRef}
\alias{.pickRef}
\title{Secret function that does the magic for pickRefSeq.}
\usage{
.pickRef(l, priorities, reduce = c("all", "first", "last"))
}
\arguments{
\item{l}{List.}

\item{priorities}{How to prioritize.}

\item{reduce}{How to reduce.}
}
\value{
List.
}
\description{
Do not use it, use \code{\link{pickRefSeq}}!
}
\note{
Hey, you found a secret function! Keep it that way!
}
\seealso{
\code{\link{pickRefSeq}}
}
\author{
Stefan McKinnon Edwards \email{stefan.hoj-edwards@agrsci.dk}
}
