% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_doc.R
\name{Edges_df}
\alias{Edges_df}
\title{get Edges dataframe for visNetwork function}
\usage{
Edges_df()
}
\value{
A data frame with egdes attributes
}
\description{
get Edges dataframe for visNetwork function
}
\examples{

\dontrun{
load(paste(.libPaths(),"/bioCancer/extdata/ListProfData.RData", sep=""))
Edges_df()
}

}

