% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_doc.R
\name{Edges_Diseases_obj}
\alias{Edges_Diseases_obj}
\title{get Edges dataframe for Gene/Disease association from geNetClassifier}
\usage{
Edges_Diseases_obj(genesclassdetails)
}
\arguments{
\item{genesclassdetails}{a dataframe from geNetClassifier function}
}
\value{
A data frame with egdes attributes
}
\description{
get Edges dataframe for Gene/Disease association from geNetClassifier
}
\examples{

load(paste(path.package("bioCancer"),"/extdata/ListProfData.RData", sep=""))
Ed_Diseases_obj <- Edges_Diseases_obj(GenesClassDetails)


}

