% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioCancer.R
\docType{data}
\name{epiGenomics}
\alias{epiGenomics}
\title{Default dataset of bioCancer}
\format{An object of class \code{data.frame} with 48 rows and 7 columns.}
\usage{
epiGenomics
}
\description{
Default dataset of bioCancer
}
\author{
Karim Mezhoud \email{kmezhoud@gmail.com}
}
\keyword{datasets}
