% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantomJS.R
\name{findPhantom}
\alias{findPhantom}
\title{Check if PhantomJS is installed. Similar to webshot}
\usage{
findPhantom()
}
\value{
Logic object
}
\description{
Check if PhantomJS is installed. Similar to webshot
}
\examples{
How <- "runManually"
\dontrun{
findPhantom()
}
}
