% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdsr.R
\name{getCancerStudies.CGDS}
\alias{getCancerStudies.CGDS}
\title{S3 method to get Cancer Studies}
\usage{
getCancerStudies.CGDS(x, ...)
}
\arguments{
\item{x}{connection object}

\item{...}{not used}
}
\description{
S3 method to get Cancer Studies
}
\examples{
# Create CGDS object
mycgds <- CGDS("http://www.cbioportal.org/")
# Get available case lists (collection of samples) for a given cancer study
mycancerstudy <- getCancerStudies.CGDS(mycgds)[2,1]

}
