% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdsr.R
\name{getCaseLists.CGDS}
\alias{getCaseLists.CGDS}
\title{S3 method to get Cases Lists}
\usage{
getCaseLists.CGDS(x, cancerStudy,...)
}
\arguments{
\item{x}{connection object}

\item{cancerStudy}{cancer study ID}

\item{...}{Not used}
}
\description{
S3 method to get Cases Lists
}
\examples{
# Create CGDS object
mycgds <- CGDS("http://www.cbioportal.org/")
# Get list of cancer studies at server
mycancerstudy <- getCancerStudies.CGDS(mycgds)[2,1]
# Get available case lists (collection of samples) for a given cancer study
mycaselist <- getCaseLists.CGDS(mycgds,mycancerstudy)[1,1]

}
