% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdsr.R
\name{getClinicalData.CGDS}
\alias{getClinicalData.CGDS}
\title{S3 method to get Clinical Data}
\usage{
getClinicalData.CGDS(x, caseList, cases, caseIdsKey, ...)
}
\arguments{
\item{x}{connection object}

\item{caseList}{A list of cases ID}

\item{cases}{A vector of case IDs}

\item{caseIdsKey}{only used by web portal}

\item{...}{not used}
}
\description{
S3 method to get Clinical Data
}
\examples{
#Create CGDS object
mycgds <- CGDS("http://www.cbioportal.org/")
# Get available case lists (collection of samples) for a given cancer study
mycancerstudy <- getCancerStudies.CGDS(mycgds)[2,1]
mycaselist <- getCaseLists.CGDS(mycgds,mycancerstudy)[1,1]
}
