% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationFuncs.R
\name{getEvidenceCodes}
\alias{getEvidenceCodes}
\title{Returns GO evidence codes.}
\usage{
getEvidenceCodes()
}
\value{
Matrix of two columns, first column with codes, second column with description of codes.
}
\description{
Returns GO evidence codes.
}
\examples{
getEvidenceCodes()
}
\references{
\code{?org.Bt.egGO}
}
\seealso{
\code{\link{pickGO}}
}
\author{
Stefan McKinnon Edwards \email{stefan.hoj-edwards@agrsci.dk}
}
