% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich.R
\name{getFreqMutData}
\alias{getFreqMutData}
\title{get mutation frequency}
\usage{
getFreqMutData(list, geneListLabel)
}
\arguments{
\item{list}{a list of data frame with mutation data. Each data frame is for one study}

\item{geneListLabel}{file name of geneList examples: "73"}
}
\value{
a data frame with mutation frequency. gene is in rows and study is in column
}
\description{
get mutation frequency
}
\examples{
cgds <- cBioPortal(
hostname = "www.cbioportal.org",
protocol = "https",
api = "/api/v2/api-docs"
)
\dontrun{
getDataByGenes( api =  cgds,
studyId = "gbm_tcga_pub",
genes = c("NF1", "TP53", "ABL1"),
by = "hugoGeneSymbol",
molecularProfileIds = "gbm_tcga_pub_mrna"
)
}
}
