% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geNetClassifier.R
\name{getGenesClassification}
\alias{getGenesClassification}
\title{get genes classification}
\usage{
getGenesClassification(checked_Studies, GeneList,
 samplesize, threshold, listGenProfs, listCases)
}
\arguments{
\item{checked_Studies}{checked studies}

\item{GeneList}{gene list}

\item{samplesize}{sample size}

\item{threshold}{p-value threshold}

\item{listGenProfs}{list of genetic profiles}

\item{listCases}{list of cases}
}
\value{
A table with genes classed by study
}
\description{
get genes classification
}
\examples{
cgds <- cBioPortal(
hostname = "www.cbioportal.org",
protocol = "https",
api = "/api/v2/api-docs"
)
\dontrun{
getDataByGenes( api =  cgds,
studyId = "gbm_tcga_pub",
genes = c("NF1", "TP53", "ABL1"),
by = "hugoGeneSymbol",
molecularProfileIds = "gbm_tcga_pub_mrna"
)
}

}
