<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> bbe4d9cf03dcd14dc250f9c2de17d39c14fb1c1b
% Please edit documentation in R/getNetClassifier_doc.R
\name{getGenesClassifier}
\alias{getGenesClassifier}
\title{Classify genes by Studies using gene expression}
\usage{
getGenesClassifier()
}
\value{
A data frame with ranking genes.
}
\description{
Classify genes by Studies using gene expression
}
\examples{
\dontrun{
load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
getGenesClassifier()
}
<<<<<<< HEAD

=======
>>>>>>> bbe4d9cf03dcd14dc250f9c2de17d39c14fb1c1b
}

