% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdsr.R
\name{getGeneticProfiles.CGDS}
\alias{getGeneticProfiles.CGDS}
\title{S3 method to get Genetic Profiles}
\usage{
getGeneticProfiles.CGDS(x, cancerStudy, ...)
}
\arguments{
\item{x}{connection object}

\item{cancerStudy}{cancer study ID}

\item{...}{not used}
}
\description{
S3 method to get Genetic Profiles
}
\examples{
# Create CGDS object
mycgds <- CGDS("http://www.cbioportal.org/")
# Get list of cancer studies at server
mycancerstudy <- getCancerStudies.CGDS(mycgds)[2,1]
# Get available case lists (collection of samples) for a given cancer study
mycaselist <- getCaseLists.CGDS(mycgds,mycancerstudy)[1,1]
# Get available genetic profiles
mygeneticprofile <- getGeneticProfiles.CGDS(mycgds,mycancerstudy)[1,1]
# Get data slices for a specified list of genes, genetic profile and case list
myProfileData <- getProfileData.CGDS(mycgds,c('BRCA1','BRCA2'),mygeneticprofile,mycaselist)

}
