% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geNetClassifier.R
\name{getList_Cases}
\alias{getList_Cases}
\title{get list of cases of each selected study in Classifier panel}
\usage{
getList_Cases(checked_Studies)
}
\arguments{
\item{checked_Studies}{checked studies}
}
\value{
A list of cases
}
\description{
get list of cases of each selected study in Classifier panel
}
\examples{
cgds <- cBioPortal(
hostname = "www.cbioportal.org",
protocol = "https",
api = "/api/v2/api-docs"
)
\dontrun{
getDataByGenes( api =  cgds,
studyId = "gbm_tcga_pub",
genes = c("NF1", "TP53", "ABL1"),
by = "hugoGeneSymbol",
molecularProfileIds = "gbm_tcga_pub_mrna"
)
}

}
