% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geNetClassifier.R
\name{getList_GenProfs}
\alias{getList_GenProfs}
\title{get list of genetic profiles of each selected study in Classifier panel}
\usage{
getList_GenProfs(checked_Studies)
}
\arguments{
\item{checked_Studies}{checked studies}
}
\value{
listes of genetics profiles
}
\description{
get list of genetic profiles of each selected study in Classifier panel
}
\examples{
\dontrun{
cgds <- CGDS("http://www.cbioportal.org/")
listStudies <-  getCancerStudies(cgds)
listGenProfs <- getList_GenProfs(listStudies[1:3])
}

}
