% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListProfData_doc.R
\name{getListProfData}
\alias{getListProfData}
\title{get list of data frame with profiles data (CNA,mRNA, Methylation, Mutation...)}
\usage{
getListProfData(panel)
}
\arguments{
\item{panel}{Panel name (string) in which Studies are selected. There are two panels ("Circomics" or "Reactome")}
}
\value{
A LIST of a list data frame. Each LIST is related to profiles data (CNA, mRNA, Methylation, Mutation, miRNA, RPPA).
        each list of data frame is related to studies.
}
\description{
get list of data frame with profiles data (CNA,mRNA, Methylation, Mutation...)
}
\details{
See \url{https://github.com/kmezhoud/bioCancer/wiki}
}
\examples{
example <- "runManually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
##load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
StudiesIDCircos <- c("luad_tcga_pub","blca_tcga_pub")
ListProfData <- getListProfData(panel= "Circomics")
}


}

