% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMegaProfData.R
\name{getMegaProfData}
\alias{getMegaProfData}
\title{Search and get genetic profiles (CNA,mRNA, Methylation, Mutation...) of gene list upper than 500}
\usage{
getMegaProfData(MegaGeneList, Study, GenProf, Case, Class)
}
\arguments{
\item{MegaGeneList}{A list of genes upper than 500}

\item{Study}{Study ID}

\item{GenProf}{genetic profile reference}

\item{Case}{Case reference}

\item{Class}{indicates the panel ProfData or Mutdata}
}
\value{
A data frame with Genetic profile
}
\description{
Search and get genetic profiles (CNA,mRNA, Methylation, Mutation...) of gene list upper than 500
}
\details{
See \url{https://github.com/kmezhoud/bioCancer/wiki}
}
\examples{
cgds <- cBioPortal(
hostname = "www.cbioportal.org",
protocol = "https",
api = "/api/v2/api-docs"
)
\dontrun{
getDataByGenes( api =  cgds,
studyId = "gbm_tcga_pub",
genes = c("NF1", "TP53", "ABL1"),
by = "hugoGeneSymbol",
molecularProfileIds = "gbm_tcga_pub_mrna"
)
}
}
