% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdsr.R
\name{getProfileData.CGDS}
\alias{getProfileData.CGDS}
\title{S3 method to get Profile Data}
\usage{
getProfileData.CGDS(x, genes, geneticProfiles, caseList, cases, caseIdsKey, ...)
}
\arguments{
\item{x}{connection object}

\item{genes}{A genes list}

\item{geneticProfiles}{A genetic Profile ID}

\item{caseList}{A cases list ID}

\item{cases}{A vector of cases ID}

\item{caseIdsKey}{Only used by web portal}

\item{...}{not used}
}
\description{
S3 method to get Profile Data
}
\examples{
# Create CGDS object
mycgds <- CGDS("http://www.cbioportal.org/")
# Get list of cancer studies at server
mycancerstudy <- getCancerStudies.CGDS(mycgds)[2,1]
# Get available case lists (collection of samples) for a given cancer study
mycaselist <- getCaseLists.CGDS(mycgds,mycancerstudy)[1,1]
# Get available genetic profiles
mygeneticprofile <- getGeneticProfiles.CGDS(mycgds,mycancerstudy)[1,1]
# Get data slices for a specified list of genes, genetic profile and case list
myProfileData <- getProfileData.CGDS(mycgds,c('BRCA1','BRCA2'),mygeneticprofile,mycaselist)
# Get data slice for a single gene
mysigneProfileData <- getProfileData.CGDS(mycgds,'HMGA2',mygeneticprofile,mycaselist)

}
