% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich.R
\name{getSequensed_SampleSize}
\alias{getSequensed_SampleSize}
\title{get samples size of sequensed genes}
\usage{
getSequensed_SampleSize(StudyID)
}
\arguments{
\item{StudyID}{Study reference using cgdsr index}
}
\value{
dataframe with sample size for each selected study.
}
\description{
get samples size of sequensed genes
}
