% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabologram.R
\name{metabologram}
\alias{metabologram}
\title{Circular plot of hierarchital data of genetic profile.}
\usage{
metabologram(treeData = sampleMetabologramData, width=600, height=600, main=title, showLegend = TRUE, fontSize = 10, legendBreaks=c("NA","Min","Negative", "0", "Positive", "Max"), legendColors=c("black","blue","cyan","white","yellow","red") , legendText="Legend")
}
\arguments{
\item{treeData}{A hierarchical tree data as in example}

\item{width}{600}

\item{height}{600}

\item{main}{Title}

\item{showLegend}{FALSE}

\item{legendColors}{NULL}

\item{fontSize}{12}

\item{legendText}{Legend}

\item{legenBreacks}{NULL}
}
\value{
A circular layout with genetic profile.
}
\description{
Circular plot of hierarchital data of genetic profile.
}
\examples{
\dontrun{
library(bioCancer)
bioCancer::metabologram(treeData = sampleMetabologramData, width=600, height=600, main=title, showLegend = TRUE, fontSize = 10, legendBreaks=c("NA","Min","Negative", "0", "Positive", "Max"), legendColors=c("black","blue","cyan","white","yellow","red") , legendText="Legend")
}

}
\seealso{
https://github.com/armish/metabologram
}

