% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabologram.R
\name{metabologramOutput}
\alias{metabologramOutput}
\title{Widget output function for use in Shiny}
\usage{
metabologramOutput(outputId, width = 600, height = 500)
}
\arguments{
\item{outputId}{id}

\item{width}{600}

\item{height}{600}
}
\value{
A circular plot with genetic profile in Shiny App.
}
\description{
Widget output function for use in Shiny
}
\examples{
\dontrun{
library(bioCancer)
bioCancer::metabologram(treeData = sampleMetabologramData)
}
}
