% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_doc.R
\name{Mutation_obj}
\alias{Mutation_obj}
\title{Atribute mutation frequency to nodes}
\usage{
Mutation_obj(list,threshold)
}
\arguments{
\item{list}{A list of data frame with mutation data. Each data frame to study}

\item{threshold}{Rate of cases (patients) having mutation (0-1).}
}
\value{
A dat frame with mutation frequency. Ech column corresponds to a study.
}
\description{
Atribute mutation frequency to nodes
}
\examples{
example <- "runManually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
##load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
Mut_obj <- Mutation_obj(ListMutData_bkp, 0.2)
}

}

