% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_doc.R
\name{Node_df}
\alias{Node_df}
\title{Data frame with Nodes id}
\usage{
Node_df(genelist)
}
\arguments{
\item{genelist}{a vector of genes}
}
\value{
A data frame with nodes id
}
\description{
Data frame with Nodes id
}
\examples{
load(paste(path.package("bioCancer"),"/extdata/ListProfData.RData", sep=""))
GeneFreq <- Node_obj(GeneList)

}

