% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_doc.R
\name{Node_Diseases_obj}
\alias{Node_Diseases_obj}
\title{Attributes color and shape to Nodes of Diseases}
\usage{
Node_Diseases_obj(genesclassdetails)
}
\arguments{
\item{genesclassdetails}{a dataframe from geNetClassifier function}
}
\value{
A data frame with nodes Shapes and colors
}
\description{
Attributes color and shape to Nodes of Diseases
}
\examples{
example <- "run manually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
Node_Diseases_df <- Node_Diseases_obj(genesclassdetails= GenesClassDetails)
}
}

