% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome.R
\name{Node_obj_CNA_ProfData}
\alias{Node_obj_CNA_ProfData}
\title{Attribute CNA data to node border}
\usage{
Node_obj_CNA_ProfData(list)
}
\arguments{
\item{list}{A list of data frame with CNA data. Each data frame corresponds to a study.}
}
\value{
A data frame with node border attributes
}
\description{
Attribute CNA data to node border
}
\examples{
cgds <- cBioPortal(
hostname = "www.cbioportal.org",
protocol = "https",
api = "/api/v2/api-docs"
)
\dontrun{
getDataByGenes( api =  cgds,
studyId = "gbm_tcga_pub",
genes = c("NF1", "TP53", "ABL1"),
by = "hugoGeneSymbol",
molecularProfileIds = "gbm_tcga_pub_mrna"
)
}
}
