% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_doc.R
\name{Node_obj_FreqIn}
\alias{Node_obj_FreqIn}
\title{Attribute interaction frequency to node size}
\usage{
Node_obj_FreqIn(GeneList)
}
\arguments{
\item{GeneList}{A list of gene symbol}
}
\value{
A data frame with node attributes
}
\description{
Attribute interaction frequency to node size
}
\examples{
example <- "runManually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
 ##load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
 ReactomeFI <- readRDS(paste0(system.file(package = "bioCancer"),"/extdata/DisGeNet.RDS"))
 node_obj <- Node_obj_FreqIn(GeneList)
}


}

