% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome.R
\name{Node_obj_Met_ProfData}
\alias{Node_obj_Met_ProfData}
\title{Attribute gene Methylation to Nodes}
\usage{
Node_obj_Met_ProfData(list, type, threshold)
}
\arguments{
\item{list}{a list of data frame with methylation data}

\item{type}{HM450 or HM27}

\item{threshold}{the Rate cases (patients) that have a silencing genes by methylation}
}
\value{
a data frame with node shape attributes
}
\description{
Attribute gene Methylation to Nodes
}
\examples{
cgds <- cBioPortal(
hostname = "www.cbioportal.org",
protocol = "https",
api = "/api/v2/api-docs"
)
\dontrun{
getDataByGenes( api =  cgds,
studyId = "gbm_tcga_pub",
genes = c("NF1", "TP53", "ABL1"),
by = "hugoGeneSymbol",
molecularProfileIds = "gbm_tcga_pub_mrna"
)
}

}
