% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_doc.R
\name{Node_obj_mRNA_Classifier}
\alias{Node_obj_mRNA_Classifier}
\title{Atrribute genes expression to color nodes}
\usage{
Node_obj_mRNA_Classifier(GeneList,genesclassdetails)
}
\arguments{
\item{GeneList}{A gene list.}

\item{genesclassdetails}{A dataframe with genes classes and genes expression.}
}
\value{
A data frame with node color attributes
}
\description{
Atrribute genes expression to color nodes
}
\examples{
example <- "runManually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
##load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
ReactomeFI <- readRDS(paste0(system.file(package = "bioCancer"),"/extdata/DisGeNet.RDS"))
node_obj <- Node_obj_mRNA_Classifier(GeneList, GenesClassDetails)
}

}

