% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdsr.R
\name{processURL.CGDS}
\alias{processURL.CGDS}
\title{S3 method to process URL}
\usage{
processURL.CGDS(x, url, force.comment.char.blank, ...)
}
\arguments{
\item{x}{A connection object}

\item{url}{URL}

\item{force.comment.char.blank}{a boolean param to force comment}

\item{...}{not used}
}
\description{
These methods should not be invoked by the user.
}
