% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationFuncs.R
\name{removeNAs}
\alias{removeNAs}
\title{Removes entries equal \code{NA} from list or vector}
\usage{
removeNAs(l)
}
\arguments{
\item{l}{Vector or list.}
}
\description{
Removes entries equal \code{NA}, but not mixed entries containing, amongst others, \code{NA}.
Good for use after \code{\link{mapLists}} that might return entries equal \code{NA}.
}
\examples{
removeNAs(list('a'=NA, 'b'=c(NA, 'B'), 'c'='C'))
}
\author{
Stefan McKinnon Edwards \email{stefan.hoj-edwards@agrsci.dk}
}
