% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffeewheel.R
\name{renderCoffeewheel}
\alias{renderCoffeewheel}
\title{Widget render function for use in Shiny}
\usage{
renderCoffeewheel(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{id}

\item{env}{parent.frame}

\item{quoted}{FALSE}
}
\value{
A circular layout with genetic profile in Shiny App.
}
\description{
Widget render function for use in Shiny
}
\examples{

example <- "runManually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
##load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
bioCancer::coffeewheel(treeData = sampleWheelData)
}
}

