% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabologram.R
\name{renderMetabologram}
\alias{renderMetabologram}
\title{Widget render function for use in Shiny}
\usage{
renderMetabologram(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{expression}

\item{env}{parent.frame()}

\item{quoted}{FALSE}
}
\value{
A circular plot with genetic profile in Shiny App.
}
\description{
Widget render function for use in Shiny
}
\examples{
\dontrun{
library(bioCancer)
bioCancer::metabologram(treeData = sampleMetabologramData)
}
}
