% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircomics_doc.R
\name{reStrColorGene}
\alias{reStrColorGene}
\title{Restructure the list of color attributed to the genes in every dimenssion for every studies}
\usage{
reStrColorGene(df)
}
\arguments{
\item{df}{data frame with colors attributed to the genes}
}
\value{
Hierarchical color attribute: gene > color
}
\description{
Restructure the list of color attributed to the genes in every dimenssion for every studies
}
\examples{
example <- "runManually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
 ##load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
Children <- reStrColorGene(ListProfData$Expression$luad_tcga_pub)
}
}

