% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich.R
\name{reStrDimension}
\alias{reStrDimension}
\title{Restructure the list of color attributed to the genes in every study for every dimensions}
\usage{
reStrDimension(LIST)
}
\arguments{
\item{LIST}{list of  hierarchical dimensions}
}
\value{
Hierarchical structure of: Study > dimensions > gene > color
}
\description{
Restructure the list of color attributed to the genes in every study for every dimensions
}
\examples{
cgds <- cBioPortal(
hostname = "www.cbioportal.org",
protocol = "https",
api = "/api/v2/api-docs"
)
\dontrun{
getDataByGenes( api =  cgds,
studyId = "gbm_tcga_pub",
genes = c("NF1", "TP53", "ABL1"),
by = "hugoGeneSymbol",
molecularProfileIds = "gbm_tcga_pub_mrna"
)
}
}
