% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircomics_doc.R
\name{reStrDisease}
\alias{reStrDisease}
\title{Restructure the list of color attributed to the genes in every disease}
\usage{
reStrDisease(List)
}
\arguments{
\item{List}{of data frame with color attributes}
}
\value{
Hierarchy of dimensions in the same study: dimensions > gene > color
}
\description{
Restructure the list of color attributed to the genes in every disease
}
\examples{
example <- "runManually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
##load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
Child <- reStrDisease(ListProfData$Expression)
}


}

