% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffeewheel.R
\docType{data}
\name{sampleWheelData}
\alias{sampleWheelData}
\title{Sample data for wheel initialization}
\format{An object of class \code{list} of length 3.}
\usage{
sampleWheelData
}
\value{
A list of hierarchical data
}
\description{
Sample data for wheel initialization
}
\examples{

example <- "runManually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
##load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
bioCancer::coffeewheel(treeData = sampleWheelData)
}
}
\keyword{datasets}

