% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdsr.R
\name{setVerbose.CGDS}
\alias{setVerbose.CGDS}
\title{S3 method to set verbose}
\usage{
setVerbose.CGDS(x, verbose, ...)
}
\arguments{
\item{x}{A connection object}

\item{verbose}{Activate verbose logging (boolean)}

\item{...}{not used}
}
\description{
Sets verbose logging level for CGDS function calls.
}
\examples{
# Create CGDS object
mycgds <- CGDS("http://www.cbioportal.org/")
# Activate verbose logging
setVerbose.CGDS(mycgds, TRUE)

}
