% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_doc.R
\name{Studies_obj}
\alias{Studies_obj}
\title{get object for grViz. Link Studies to genes}
\usage{
Studies_obj(df)
}
\arguments{
\item{df}{data frame with gene classes}
}
\value{
grViz object. a data frame with Study attributes
}
\description{
get object for grViz. Link Studies to genes
}
\examples{
\dontrun{
load(paste(.libPaths(),"/bioCancer/extdata/ListProfData.RData", sep=""))
Sd_obj <- Studies_obj(GenesClassDetails)
}
}

