% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchButton.R
\name{switchButton}
\alias{switchButton}
\title{A function to change the Original checkbox of rshiny
into a nice true/false or on/off switch button
No javascript involved. Only CSS code.}
\usage{
switchButton(inputId, label = NULL, value = FALSE, col = "GB",
  type = "TF")
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value (TRUE or FALSE).}

\item{col}{Color set of the switch button. Choose between "GB" (Grey-Blue) and "RG" (Red-Green)}

\item{type}{Text type of the button. Choose between "TF" (TRUE - FALSE), "OO" (ON - OFF) or leave empty for no text.}
}
\description{
To be used with CSS script 'button.css' stored in a 'www' folder in your Shiny app folder
}
