% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdsr.R
\name{test.CGDS}
\alias{test.CGDS}
\title{S3 method to test cBioPortal connection}
\usage{
\method{test}{CGDS}(x, ...)
}
\arguments{
\item{x}{connection object}

\item{...}{not used}
}
\description{
S3 method to test cBioPortal connection
}
