% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioCancer.R
\docType{data}
\name{user_CNA}
\alias{user_CNA}
\title{Example of Copy Number Alteration (CNA) dataset}
\format{An object of class \code{data.frame} with 579 rows and 13 columns.}
\usage{
user_CNA
}
\description{
Example of Copy Number Alteration (CNA) dataset
}
\author{
Karim Mezhoud \email{kmezhoud@gmail.com}
}
\keyword{datasets}
