% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioCancer.R
\docType{data}
\name{user_MetHM27}
\alias{user_MetHM27}
\title{Example of Methylation HM27 dataset}
\format{An object of class \code{data.frame} with 600 rows and 13 columns.}
\usage{
user_MetHM27
}
\description{
Example of Methylation HM27 dataset
}
\author{
Karim Mezhoud \email{kmezhoud@gmail.com}
}
\keyword{datasets}
