% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioCancer.R
\docType{data}
\name{user_MetHM450}
\alias{user_MetHM450}
\title{Example of Methylation HM450 dataset}
\format{An object of class \code{data.frame} with 10 rows and 13 columns.}
\usage{
user_MetHM450
}
\description{
Example of Methylation HM450 dataset
}
\author{
Karim Mezhoud \email{kmezhoud@gmail.com}
}
\keyword{datasets}
