% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioCancer.R
\docType{data}
\name{user_mRNA}
\alias{user_mRNA}
\title{Example of mRNA expression dataset}
\format{
An object of class \code{data.frame} with 307 rows and 13 columns.
}
\usage{
user_mRNA
}
\description{
Example of mRNA expression dataset
}
\author{
Karim Mezhoud \email{kmezhoud@gmail.com}
}
\keyword{datasets}
