% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioCancer.R
\docType{data}
\name{user_Mut}
\alias{user_Mut}
\title{Example of Mutation dataset}
\format{An object of class \code{data.frame} with 37 rows and 23 columns.}
\usage{
user_Mut
}
\description{
Example of Mutation dataset
}
\author{
Karim Mezhoud \email{kmezhoud@gmail.com}
}
\keyword{datasets}
