% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withMathJaxR.R
\name{withMathJax}
\alias{withMathJax}
\title{Use MathJax off-line in Shiny apps}
\usage{
withMathJax(...)
}
\arguments{
\item{...}{HTML input to be processed by MathJax}
}
\value{
Processed HTML
}
\description{
Use MathJax off-line in Shiny apps
}
\details{
Code adapted form shiny::withMathJax
}
\examples{
\dontrun{
withMathJax("Some math here $$\\\\alpha+\\\\beta$$")
}
}

