% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withMathJaxR.R
\name{withMathJaxR}
\alias{withMathJaxR}
\title{Use MathJax off-line in Shiny apps}
\usage{
withMathJaxR(...)
}
\arguments{
\item{...}{HTML input to be processed by MathJax}
}
\value{
Processed HTML
}
\description{
Use MathJax off-line in Shiny apps
}
\details{
Code adapted from shiny::withMathJax
}
\examples{
bioCancer::withMathJax("Some math here $$\\\\alpha+\\\\beta$$")

}

