% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich.R
\name{attriColorGene}
\alias{attriColorGene}
\title{Attribute Color to Gene}
\usage{
attriColorGene(df)
}
\arguments{
\item{df}{data frame with mRNA or CNA or mutation frequency or methylation (numeric).}
}
\value{
A list colors for every gene
}
\description{
Attribute Color to Gene
}
\examples{
cgds <- cBioPortal(
hostname = "www.cbioportal.org",
protocol = "https",
api = "/api/v2/api-docs"
)
\dontrun{
getDataByGenes( api =  cgds,
studyId = "gbm_tcga_pub",
genes = c("NF1", "TP53", "ABL1"),
by = "hugoGeneSymbol",
molecularProfileIds = "gbm_tcga_pub_mrna"
)
}
}
