% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircomics_doc.R
\name{attriColorGene}
\alias{attriColorGene}
\title{Attribute Color to Gene}
\usage{
attriColorGene(df)
}
\arguments{
\item{df}{data frame with mRNA or CNA or mutation frequency or methylation (numeric).}
}
\value{
A list colors for every gene
}
\description{
Attribute Color to Gene
}
\examples{
example <- "runManuallly"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
 ###load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
colorsls <- attriColorGene(ListProfData$Expression$luad_tcga_pub)
}
}

