% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircomics_doc.R
\name{attriColorGene}
\alias{attriColorGene}
\title{Attribute Color to Gene}
\usage{
attriColorGene(df)
}
\arguments{
\item{df}{data frame with mRNA or CNA or mutation frequency or methylation (numeric).}
}
\value{
A list colors for every gene
}
\description{
Attribute Color to Gene
}
\examples{
\dontrun{
load(paste(.libPaths(),"/bioCancer/extdata/ListProfData.RData", sep=""))
colorsls <- attriColorGene(ListProfData$Expression$luad_tcga_pub)
}
}

