% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich.R
\name{attriColorGene}
\alias{attriColorGene}
\title{Attribute Color to Gene}
\usage{
attriColorGene(df)
}
\arguments{
\item{df}{data frame with mRNA or CNA or mutation frequency or methylation (numeric).}
}
\value{
A list colors for every gene
}
\description{
Attribute Color to Gene
}
\examples{
\dontrun{
cgds <- CGDS("http://www.cbioportal.org/")
geneList <- whichGeneList("73")
ProfData <- getProfileData.CGDS(cgds,
 geneList, "gbm_tcga_pub_mrna", "gbm_tcga_pub_all")
rownames(ProfData) <- NULL
clr <- attriColorGene(ProfData)
}
}
