% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich.R
\name{attriColorValue}
\alias{attriColorValue}
\title{Attribute Color to Value}
\usage{
attriColorValue(Value, df, colors=c(a,b,c),feet)
}
\arguments{
\item{Value}{integer}

\item{df}{data frame with numeric values}

\item{colors}{a vector of 5 colors}

\item{feet}{the interval between two successive colors in the palette (0.1)}
}
\value{
Hex Color Code
}
\description{
Attribute Color to Value
}
\examples{
cgds <- cBioPortal(
hostname = "www.cbioportal.org",
protocol = "https",
api = "/api/v2/api-docs"
)
\dontrun{
getDataByGenes( api =  cgds,
studyId = "gbm_tcga_pub",
genes = c("NF1", "TP53", "ABL1"),
by = "hugoGeneSymbol",
molecularProfileIds = "gbm_tcga_pub_mrna"
)
}

}
