<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> bbe4d9cf03dcd14dc250f9c2de17d39c14fb1c1b
% Please edit documentation in R/Reactome_doc.R
\name{attriColorVector}
\alias{attriColorVector}
\title{Attribute colorto a vector of numeric values}
\usage{
attriColorVector(Value, vector, colors = c(a, b, c), feet)
}
\arguments{
\item{Value}{numeric}

\item{vector}{A vector of numeric data}

\item{colors}{3 colors}

\item{feet}{An interval between two numeric value needed to change the color}
}
\value{
A vetor of colors
}
\description{
Attribute colorto a vector of numeric values
}
\examples{
\dontrun{
load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
sapply(GenesClassDetails$exprsMeanDiff, function(x) as.character(attriColorVector(x,GenesClassDetails$exprsMeanDiff ,colors=c("blue","white","red"), feet=1)))
}
}

