% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome.R
\name{attriColorVector}
\alias{attriColorVector}
\title{Attribute color to a vector of numeric values}
\usage{
attriColorVector(Value, vector, colors=c(a,b,c),feet)
}
\arguments{
\item{Value}{numeric}

\item{vector}{A vector of numeric data}

\item{colors}{3 colors}

\item{feet}{An interval between two numeric value needed to change the color}
}
\value{
A vetor of colors
}
\description{
Attribute color to a vector of numeric values
}
\examples{
cgds <- CGDS("http://www.cbioportal.org/public-portal/")
\dontrun{
geneList <- whichGeneList("73")
ProfData <- getProfileData(cgds,
 geneList, "gbm_tcga_pub_mrna", "gbm_tcga_pub_all")
rownames(ProfData) <- NULL
clrVec <- attriColorVector(1.2,
ProfData[1,],
 colors = c("blue", "white","red"),
  feet=1)
}
}
