% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_doc.R
\name{attriColorVector}
\alias{attriColorVector}
\title{Attribute color to a vector of numeric values}
\usage{
attriColorVector(Value, vector, colors = c(a, b, c), feet)
}
\arguments{
\item{Value}{numeric}

\item{vector}{A vector of numeric data}

\item{colors}{3 colors}

\item{feet}{An interval between two numeric value needed to change the color}
}
\value{
A vetor of colors
}
\description{
Attribute color to a vector of numeric values
}
\examples{
load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
sapply(GenesClassDetails$exprsMeanDiff,
function(x) as.character(attriColorVector(x,
GenesClassDetails$exprsMeanDiff ,
colors=c("blue","white","red"),
 feet=1)))

}

