% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome.R
\name{attriShape2Gene}
\alias{attriShape2Gene}
\title{Attribute shape to nodes}
\usage{
attriShape2Gene(gene, genelist)
}
\arguments{
\item{gene}{Gene symbol}

\item{genelist}{Gene list}
}
\value{
A character "BRCA1[shape = 'circle', "
}
\description{
Attribute shape to nodes
}
\examples{
how <- "runManually"
\dontrun{
GeneList <- whichGeneList("73")
attriShape2Gene("P53", GeneList)
attriShape2Gene("GML",GeneList)
}

}

