<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> bbe4d9cf03dcd14dc250f9c2de17d39c14fb1c1b
% Please edit documentation in R/Reactome_doc.R
\name{attriShape2Gene}
\alias{attriShape2Gene}
\title{Attribute shape to nodes}
\usage{
attriShape2Gene(gene, genelist)
}
\arguments{
\item{gene}{Gene symbol}

\item{genelist}{Gene list}
}
\value{
A character "BRCA1[shape = 'circle', "
}
\description{
Attribute shape to nodes
}
\examples{
\dontrun{
load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
Shape_object <- attriShape2Gene("BRCA1", GeneList)
}
}

