% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_doc.R
\name{attriShape2Gene}
\alias{attriShape2Gene}
\title{Attribute shape to nodes}
\usage{
attriShape2Gene(gene, genelist)
}
\arguments{
\item{gene}{Gene symbol}

\item{genelist}{Gene list}
}
\value{
A character "BRCA1[shape = 'circle', "
}
\description{
Attribute shape to nodes
}
\examples{
example <- "runManually"
\dontrun{
load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
Shape_object <- attriShape2Gene("BRCA1", GeneList)
}

}

