% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioCancer.R
\docType{package}
\name{bioCancer}
\alias{bioCancer}
\alias{bioCancer-package}
\title{Launch bioCancer with default browser}
\value{
web page of bioCancer Shiny App
}
\description{
Launch bioCancer with default browser
}
\examples{
\dontrun{
bioCancer::bioCancer()
}

bioCancer

}

