% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioCancer.R
\docType{package}
\name{bioCancer}
\alias{-package}
\alias{bioCancer}
\title{Launch bioCancer with default browser}
\usage{
bioCancer()
}
\value{
web page of bioCancer Shiny App
}
\description{
The Main function to run bioCancer App
}
\examples{
ShinyApp <-  1
\dontrun{
bioCancer()
}
}
