<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> bbe4d9cf03dcd14dc250f9c2de17d39c14fb1c1b
% Please edit documentation in R/bioCancer.R
\name{bioCancer}
\alias{bioCancer}
\title{Launch bioCancer with default browser}
\usage{
bioCancer()
}
\value{
web page of bioCancer Shiny App
}
\description{
Launch bioCancer with default browser
}
\examples{
\dontrun{
bioCancer()
}
}

