% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bioCancer.R
\name{bioCancer}
\alias{bioCancer}
\title{Launch bioCancer with default browser}
\usage{
bioCancer()
}
\value{
web page of bioCancer Shiny App
}
\description{
Launch bioCancer with default browser
}
\examples{
\dontrun{
bioCancer()
}
}

