% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircomics_doc.R
\name{checkDimensions}
\alias{checkDimensions}
\title{Chech wich Cases and genetic profiles are available for every seleted study}
\usage{
checkDimensions(panel)
}
\arguments{
\item{panel}{panel can take to strings 'Circomics' or 'Reactome'}
}
\value{
A data frame with two column (Cases, Genetic profiles). Every row has a dimension (CNA, mRNA...).
The data frame is filled with yes/no response.
}
\description{
Chech wich Cases and genetic profiles are available for every seleted study
}
\examples{
example <- "runManually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
##load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
df <- checkDimensions(panel='Reactome')
}

}

