% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich.R
\name{checkDimensions}
\alias{checkDimensions}
\title{Chech wich Cases and genetic profiles are available for every seleted study}
\usage{
checkDimensions(panel,StudyID)
}
\arguments{
\item{panel}{panel can take to strings 'Circomics' or 'Networking'}

\item{StudyID}{Study reference using cgdsr index}
}
\value{
A data frame with two column (Cases, Genetic profiles). Every row has a dimension (CNA, mRNA...).
The data frame is filled with yes/no response.
}
\description{
Chech wich Cases and genetic profiles are available for every seleted study
}
\examples{
cgds <- CGDS("http://www.cbioportal.org/")
\dontrun{
df <- checkDimensions(panel='Networking', StudyID= "gbm_tcga_pub")
}
}
