% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffeewheel.R
\name{coffeewheelOutput}
\alias{coffeewheelOutput}
\title{Widget output function for use in Shiny}
\usage{
coffeewheel(treeData = sampleWheelData)
}
\arguments{
\item{width}{600}

\item{height}{600}

\item{output}{id}
}
\value{
A circular layout with genetic profile in Shiny App.
}
\description{
Widget output function for use in Shiny
}
\examples{

bioCancer::coffeewheel(treeData = sampleWheelData)

}

