% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffeewheel.R
\name{coffeewheelOutput}
\alias{coffeewheelOutput}
\title{Widget output function for use in Shiny}
\usage{
coffeewheelOutput(outputId, width=700, height=700)
}
\arguments{
\item{outputId}{id}

\item{width}{700}

\item{height}{700}
}
\value{
A circular layout with genetic profile in Shiny App.
}
\description{
Widget output function for use in Shiny
}
\examples{
How <- "runManually"
\dontrun{
coffeewheel(treeData = sampleWheelData)
}
}

