% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffeewheel.R
\name{coffeewheelOutput}
\alias{coffeewheelOutput}
\title{Widget output function for use in Shiny}
\usage{
coffeewheelOutput(outputId, width=700, height=700)
}
\arguments{
\item{outputId}{id}

\item{width}{700}

\item{height}{700}
}
\value{
A circular layout with genetic profile in Shiny App.
}
\description{
Widget output function for use in Shiny
}
\examples{

example <- "runManually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
##load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
 bioCancer::coffeewheel(treeData = sampleWheelData)
}
}

