% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portal.R
\name{displayTable}
\alias{displayTable}
\title{Display dataframe in table using DT package}
\usage{
displayTable(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
A table
}
\description{
Display dataframe in table using DT package
}
\examples{
\dontrun{
session <- NULL
cgds <- CGDS("http://www.cbioportal.org/")
Studies<- getCancerStudies.CGDS(cgds)
displayTable(Studies)
}
}
