% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_doc.R
\name{Edges_df}
\alias{Edges_df}
\title{get Edges dataframe for visNetwork function}
\usage{
Edges_df()
}
\value{
A data frame with egdes attributes
}
\description{
get Edges dataframe for visNetwork function
}
\examples{

load(paste(path.package("bioCancer"),"/extdata/ListProfData.RData", sep=""))
Ed_obj <- Edges_obj()


}

