% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_doc.R
\name{Edges_df}
\alias{Edges_df}
\title{get Edges dataframe for visNetwork function}
\usage{
Edges_df()
}
\value{
A data frame with egdes attributes
}
\description{
get Edges dataframe for visNetwork function
}
\examples{
example <- "run manually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
##load(paste(path.package("bioCancer"),"/extdata/ListProfData.RData", sep=""))
Edges_df()
}


}

