% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_doc.R
\name{Edges_Diseases_obj}
\alias{Edges_Diseases_obj}
\title{get Edges dataframe for Gene/Disease association from geNetClassifier}
\usage{
Edges_Diseases_obj(genesclassdetails)
}
\arguments{
\item{genesclassdetails}{a dataframe from geNetClassifier}
}
\value{
A data frame with egdes attributes
}
\description{
get Edges dataframe for Gene/Disease association from geNetClassifier
}
\examples{
\dontrun{
load(paste(.libPaths(),"/bioCancer/extdata/ListProfData.RData", sep=""))
Ed_Diseases_obj <- Edges_Diseases_obj(genesclassdetails=GenesClassDetails)
}

}

