% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_doc.R
\name{Edges_Diseases_obj}
\alias{Edges_Diseases_obj}
\title{get Edges dataframe for Gene/Disease association from geNetClassifier}
\usage{
Edges_Diseases_obj(genesclassdetails)
}
\arguments{
\item{genesclassdetails}{a dataframe from geNetClassifier}
}
\value{
A data frame with egdes attributes
}
\description{
get Edges dataframe for Gene/Disease association from geNetClassifier
}
\examples{
example <- "run Manually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
Ed_Diseases_obj <- Edges_Diseases_obj(genesclassdetails=GenesClassDetails)
}

}

