% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_doc.R
\name{Edges_obj}
\alias{Edges_obj}
\title{get Edges object for grVis function}
\usage{
Edges_obj()
}
\value{
A data frame with egdes attributes
}
\description{
get Edges object for grVis function
}
\examples{
example <- "runManually"
\dontrun{
load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
ReactomeFI <- readRDS(paste0(system.file(package = "bioCancer"),
"/extdata/all_gene_disease_association.RDS"))
Ed_obj <- Edges_obj()
}

}

