<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> bbe4d9cf03dcd14dc250f9c2de17d39c14fb1c1b
% Please edit documentation in R/Reactome_doc.R
\name{Edges_obj}
\alias{Edges_obj}
\title{get Edges object for grVis function}
\usage{
Edges_obj()
}
\value{
A data frame with egdes attributes
}
\description{
get Edges object for grVis function
}
\examples{
\dontrun{
load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
Ed_obj <- Edges_obj()
}
<<<<<<< HEAD

=======
>>>>>>> bbe4d9cf03dcd14dc250f9c2de17d39c14fb1c1b
}

