% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_doc.R
\name{Edges_obj}
\alias{Edges_obj}
\title{get Edges object for grVis function}
\usage{
Edges_obj()
}
\value{
A data frame with egdes attributes
}
\description{
get Edges object for grVis function
}
\examples{

load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
Ed_obj <- Edges_obj()


}

