% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_doc.R
\name{getAnnoGeneSet_df}
\alias{getAnnoGeneSet_df}
\title{Gene Set enrichment using Reactome FI methods}
\usage{
getAnnoGeneSet_df(genelist, type)
}
\arguments{
\item{genelist}{vector of genes}

\item{type}{"BP", "pathway","CC", "MF"}
}
\value{
A data frame with Gene Set enrichment: BP,MF, CC, Pathway, pval, FDR
}
\description{
Gene Set enrichment using Reactome FI methods
}
\examples{
example <- "run manually"
\dontrun{
load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
getAnnoGeneSet_df(GeneList, "BP")
}
}

