% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircomics_doc.R
\name{getCoffeeWheel}
\alias{getCoffeeWheel}
\title{get Coffe Wheel for Profiles Data}
\usage{
getCoffeeWheel(ListProfData)
}
\arguments{
\item{ListProfData}{a list of data frame for each dimension ans each disease.}
}
\value{
A tree structure of listes:  dimensions > Studies > genes > colors
}
\description{
get Coffe Wheel for Profiles Data
}
\examples{
\dontrun{
load(paste(.libPaths(),"/bioCancer/extdata/ListProfData.RData", sep=""))
getCoffeeWheel(ListProfData_bkp)
}

}

