% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCircomics_doc.R
\name{getCoffeeWheel}
\alias{getCoffeeWheel}
\title{get Coffe Wheel for Profiles Data}
\usage{
getCoffeeWheel(ListProfData)
}
\arguments{
\item{ListProfData}{a list of data frame for each dimension ans each disease.}
}
\value{
A tree structure of listes:  dimensions > Studies > genes > colors
}
\description{
get Coffe Wheel for Profiles Data
}
\examples{
\dontrun{
library(CoffeeWheel)

load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
getCoffeeWhel(ListPorofData_bkp)
}
}

