% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircomics_doc.R
\name{getCoffeeWheel}
\alias{getCoffeeWheel}
\title{get Coffe Wheel for Profiles Data}
\usage{
getCoffeeWheel(ListProfData)
}
\arguments{
\item{ListProfData}{a list of data frame for each dimension ans each disease.}
}
\value{
A tree structure of listes:  dimensions > Studies > genes > colors
}
\description{
get Coffe Wheel for Profiles Data
}
\examples{
example <- "runManually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
##load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
getCoffeeWheel(ListProfData)
}

}

