<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> bbe4d9cf03dcd14dc250f9c2de17d39c14fb1c1b
% Please edit documentation in R/getCircomics_doc.R
\name{getFreqMutData}
\alias{getFreqMutData}
\title{get mutation frequency}
\usage{
getFreqMutData(list)
}
\arguments{
\item{list}{a list of data frame with mutation data. Each data frame is for one study}
}
\value{
a data frame with mutation frequency. gene is in rows and study is in column
}
\description{
get mutation frequency
}
\examples{
\dontrun{
load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
Freq_DfMutData <- getFreqMutData(ListMutData_bkp)
}
}

