% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich.R
\name{getFreqMutData}
\alias{getFreqMutData}
\title{get mutation frequency}
\usage{
getFreqMutData(list, geneListLabel)
}
\arguments{
\item{list}{a list of data frame with mutation data. Each data frame is for one study}

\item{geneListLabel}{file name of geneList examples: "73"}
}
\value{
a data frame with mutation frequency. gene is in rows and study is in column
}
\description{
get mutation frequency
}
\examples{
\dontrun{
cgds <- CGDS("http://www.cbioportal.org/")
geneList <- whichGeneList("73")
r_data <- new.env()
MutData <- getMutationData(cgds,"gbm_tcga_pub_all",
 "gbm_tcga_pub_mutations", geneList )
FreqMut <- getFreqMutData(list(ls1=MutData, ls2=MutData), "73")
}
}
