% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geNetClassifier.R
\name{getGenesClassification}
\alias{getGenesClassification}
\title{get genes classification}
\usage{
getGenesClassification(checked_Studies, GeneList,
 samplesize, threshold, listGenProfs, listCases)
}
\arguments{
\item{checked_Studies}{checked studies}

\item{GeneList}{gene list}

\item{samplesize}{sample size}

\item{threshold}{p-value threshold}

\item{listGenProfs}{list of genetic profiles}

\item{listCases}{list of cases}
}
\value{
A table with genes classed by study
}
\description{
get genes classification
}
\examples{
\dontrun{
cgds <- CGDS("http://www.cbioportal.org/")
listStudies <-  getCancerStudies(cgds)
checked_Stdudies <- listStudies[3:5]
listCases <- getList_Cases(listStudies[1:3])
listGenProfs <- getList_GenProfs(listStudies[1:3])
GeneList <- c('P53', 'IFI16', 'BRCA1')
samplesize <- 50
threshold <- 0.95
table <- getGenesClassification(checked_Studies, GeneList,
samplesize  ,threshold  ,listGenProfs, listCases)
}

}
