% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetClassifier_doc.R
\name{getGenesClassifier}
\alias{getGenesClassifier}
\title{Classify genes by Studies using gene expression}
\usage{
getGenesClassifier()
}
\value{
A data frame with ranking genes.
}
\description{
Classify genes by Studies using gene expression
}
\examples{
\dontrun{
load(paste(.libPaths(),"/bioCancer/extdata/ListProfData.RData", sep=""))
getGenesClassifier()
}

}

