% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetClassifier_doc.R
\name{getGenesClassifier}
\alias{getGenesClassifier}
\title{Classify genes by Studies using gene expression}
\usage{
getGenesClassifier()
}
\value{
A data frame with ranking genes.
}
\description{
Classify genes by Studies using gene expression
}
\examples{
example <- "runManually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
##load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
getGenesClassifier()
}

}

