% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geNetClassifier.R
\name{getList_Cases}
\alias{getList_Cases}
\title{get list of cases of each selected study in Classifier panel}
\usage{
getList_Cases(checked_Studies)
}
\arguments{
\item{checked_Studies}{checked studies}
}
\value{
listes of cases
}
\description{
get list of cases of each selected study in Classifier panel
}
\examples{
cgds <- CGDS("http://www.cbioportal.org/public-portal/")
listStudies <-  cgdsr::getCancerStudies(cgds)
\dontrun{
listCases <- getList_Cases(listStudies[1:3])
}

}
