% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMegaProfData.R
\name{getMegaProfData}
\alias{getMegaProfData}
\title{search and get genetic profiles (CNA,mRNA, Methylation, Mutation...) of gene list upper than 500}
\usage{
getMegaProfData(MegaGeneList, GenProf, Case, Class)
}
\arguments{
\item{MegaGeneList}{A list of genes upper than 500}

\item{GenProf}{genetic profile reference}

\item{Case}{Case reference}

\item{Class}{indicates the panel ProfData or Mutdata}
}
\value{
A data frame with Genetic profile
}
\description{
search and get genetic profiles (CNA,mRNA, Methylation, Mutation...) of gene list upper than 500
}
\details{
See \url{https://github.com/kmezhoud/bioCancer/wiki}
}
\examples{
GeneList <- c("ALK", "JAK3", "SHC3","TP53","MYC","PARP")
\dontrun{
cgds <- CGDS("http://www.cbioportal.org/")
listCase_gbm_tcga_pub <- getCaseLists(cgds,"gbm_tcga_pub")[,1]
listGenProf_gbm_tcga_pub <- getGeneticProfiles(cgds,"gbm_tcga_pub")[,1]

ProfData_Mut <- grepRef("gbm_tcga_pub_all", listCase_gbm_tcga_pub,
 "gbm_tcga_pub_mutations", listGenProf_gbm_tcga_pub, GeneList, Mut=1)
}
}
