% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_doc.R
\name{graph_obj}
\alias{graph_obj}
\title{get graph object for grViz}
\usage{
graph_obj(NodeAttri_Reactome, NodeAttri_Classifier, NodeAttri_ProfData)
}
\arguments{
\item{NodeAttri_Reactome}{Node attribute from Reactome database ('Freq. Interaction')}

\item{NodeAttri_Classifier}{Node attributes from geNetClassifier ('mRNA','Studies','mRNA/Studies')}

\item{NodeAttri_ProfData}{Node attributes from Profiles Data ('Mutation', 'CNA', 'Met_HM450', 'Met_HM27')}
}
\value{
Object graph for grViz with Nodes and Edges attributes
}
\description{
get graph object for grViz
}
\examples{

load(paste(path.package("bioCancer"),"/extdata/ListProfData.RData", sep=""))
gr_obj <- graph_obj('Freq. Interaction', 'mRNA', 'Met_HM450')


}

