% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListProfData_doc.R
\name{grepRef}
\alias{grepRef}
\title{search and get genetic profiles (CNA,mRNA, Methylation, Mutation...)}
\usage{
ProfData <- grepRef(regex1, listRef1,regex2, listRef2, GeneList,Mut)
}
\arguments{
\item{regex1}{Case id (cancer_study_id_[mutations, cna, methylation, mrna ]).}

\item{listRef1}{A list of cases for one study.}

\item{regex2}{Genetic Profile id (cancer_study_id_[mutations, cna, methylation, mrna ]).}

\item{listRef2}{A list of Genetic Profiles for one study.}

\item{GeneList}{A list of genes}

\item{Mut}{Condition to set if the genetic profile is mutation or not (0,1)}
}
\value{
A data frame with Genetic profile
}
\description{
search and get genetic profiles (CNA,mRNA, Methylation, Mutation...)
}
\details{
See \url{https://github.com/kmezhoud/bioCancer/wiki}
}
\examples{
\dontrun{
library(bioCancer)
GeneList <- c("ALK", "JAK3", "SHC3","TP53","MYC","PARP")
cgds <- cgdsr::CGDS("http://www.cbioportal.org/public-portal/")
listCase_gbm_tcga_pub <- cgdsr::getCaseLists(cgds,"gbm_tcga_pub")[,1]
listGenProf_gbm_tcga_pub <- cgdsr::getGeneticProfiles(cgds,"gbm_tcga_pub")[,1]
ProfData_Mut <- grepRef("gbm_tcga_pub_all", listCase_gbm_tcga_pub, "gbm_tcga_pub_mutations", listGenProf_gbm_tcga_pub, GeneList, Mut=1)
}
}

