% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome.R
\name{Mutation_obj}
\alias{Mutation_obj}
\title{Atribute mutation frequency to nodes}
\usage{
Mutation_obj(list,FreqMutThreshold, geneListLabel)
}
\arguments{
\item{list}{A list of data frame with mutation data. Each data frame to study}

\item{FreqMutThreshold}{threshold Rate of cases (patients) having mutation (0-1).}

\item{geneListLabel}{file name of geneList examples: "73"}
}
\value{
A dat frame with mutation frequency. Ech column corresponds to a study.
}
\description{
Atribute mutation frequency to nodes
}
\examples{
cgds <- cBioPortal(
hostname = "www.cbioportal.org",
protocol = "https",
api = "/api/v2/api-docs"
)
\dontrun{
getDataByGenes( api =  cgds,
studyId = "gbm_tcga_pub",
genes = c("NF1", "TP53", "ABL1"),
by = "hugoGeneSymbol",
molecularProfileIds = "gbm_tcga_pub_mrna"
)
}

}
