% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome.R
\name{Mutation_obj}
\alias{Mutation_obj}
\title{Atribute mutation frequency to nodes}
\usage{
Mutation_obj(list,FreqMutThreshold, geneListLabel)
}
\arguments{
\item{list}{A list of data frame with mutation data. Each data frame to study}

\item{FreqMutThreshold}{threshold Rate of cases (patients) having mutation (0-1).}

\item{geneListLabel}{file name of geneList examples: "73"}
}
\value{
A dat frame with mutation frequency. Ech column corresponds to a study.
}
\description{
Atribute mutation frequency to nodes
}
\examples{
 cgds <- CGDS("http://www.cbioportal.org/public-portal/")
 \dontrun{
 geneList <- whichGeneList("73")
 MutData <- getMutationData(cgds,"gbm_tcga_pub_all",
 "gbm_tcga_pub_mutations", geneList )
 listMutData <- list(ls1=MutData, ls2=MutData)
 FreqMutThreshold <- 10
 r_data <- new.env()
 MutObj <- Mutation_obj(listMutData, 10, "73")
 }

}
