% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Reactome_doc.R
\name{Mutation_obj}
\alias{Mutation_obj}
\title{Atribute mutation frequency to nodes}
\usage{
Mutation_obj(list, threshold)
}
\arguments{
\item{list}{A list of data frame with mutation data. Each data frame to study}

\item{threshold}{Rate of cases (patients) having mutation (0-1).}
}
\value{
A dat frame with mutation frequency. Ech column corresponds to a study.
}
\description{
Atribute mutation frequency to nodes
}
\examples{
\dontrun{
load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
Mut_obj <- Mutation_obj(ListMutData_bkp, 0.2)
}
}

