% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_doc.R
\name{Node_df}
\alias{Node_df}
\title{Data frame with Nodes id}
\usage{
Node_df(genelist, freqIn)
}
\arguments{
\item{genelist}{a vector of genes}

\item{freqIn}{dataframe with Node interaction frequencies}
}
\value{
A data frame with nodes id
}
\description{
Data frame with Nodes id
}
\examples{

example <- "run Manually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
node_df <- Node_df(GeneList, FreqIn )
}

}

