% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_doc.R
\name{Node_df}
\alias{Node_df}
\title{Data frame with Nodes id}
\usage{
Node_df(genelist, freqIn)
}
\arguments{
\item{genelist}{a vector of genes}

\item{freqIn}{dataframe with Node interaction frequencies}
}
\value{
A data frame with nodes id
}
\description{
Data frame with Nodes id
}
\examples{

\dontrun{
load(paste(.libPaths(),"/bioCancer/extdata/ListProfData.RData", sep=""))
bioCancer::bioCancer()
}

}

