% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_doc.R
\name{Node_df_FreqIn}
\alias{Node_df_FreqIn}
\title{Attributes size to Nodes depending on number of interaction}
\usage{
Node_df_FreqIn(genelist)
}
\arguments{
\item{genelist}{a vector of genes}
}
\value{
A data frame with nodes size attributes
}
\description{
Attributes size to Nodes depending on number of interaction
}
\examples{

load(paste(path.package("bioCancer"),"/extdata/ListProfData.RData", sep=""))
GeneFreq <- Node_obj_FreqIn(GeneList)

}

