% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyNetwork.R
\name{Node_Diseases_obj}
\alias{Node_Diseases_obj}
\title{Attributes color and shape to Nodes of Diseases}
\usage{
Node_Diseases_obj(genesclassdetails)
}
\arguments{
\item{genesclassdetails}{a dataframe from geNetClassifier function}
}
\value{
A data frame with nodes Shapes and colors
}
\description{
Attributes color and shape to Nodes of Diseases
}
\examples{
GenesClassDetails <- structure(list(Genes = c("FANCF", "MLH1", "MSH2", "ATR", "PARP1",
"CHEK2", "RAD51"), ranking = c(1L, 1L, 1L, 2L, 3L, 1L, 2L), class = c("brca_tcga",
"gbm_tcga", "lihc_tcga", "lihc_tcga", "lihc_tcga", "lusc_tcga",
"lusc_tcga"), postProb = c(1, 0.99, 1, 0.99, 0.99, 1,
0.98), exprsMeanDiff = c(180, 256, -373, -268,
-1482, 258, 143), exprsUpDw = c("UP", "UP", "DOWN",
"DOWN", "DOWN", "UP", "UP")), .Names = c("Genes", "ranking",
"class", "postProb", "exprsMeanDiff", "exprsUpDw"),
class = "data.frame", row.names = c(NA,-7L))
Node_Diseases_df <- Node_Diseases_obj(genesclassdetails= GenesClassDetails)

}

