% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_doc.R
\name{Node_Diseases_obj}
\alias{Node_Diseases_obj}
\title{Attributes color and shape to Nodes of Diseases}
\usage{
Node_Diseases_obj(genesclassdetails)
}
\arguments{
\item{genesclassdetails}{a dataframe from geNetClassifier function}
}
\value{
A data frame with nodes Shapes and colors
}
\description{
Attributes color and shape to Nodes of Diseases
}
\examples{
\dontrun{
load(paste(.libPaths(),"/bioCancer/extdata/ListProfData.RData", sep=""))
Node_Diseases_df <- Node_Diseases_obj(genesclassdetails= GenesClassDetails)
}
}

