<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> bbe4d9cf03dcd14dc250f9c2de17d39c14fb1c1b
% Please edit documentation in R/Reactome_doc.R
\name{Node_obj_CNA_ProfData}
\alias{Node_obj_CNA_ProfData}
\title{Attribute CNA data to node border}
\usage{
Node_obj_CNA_ProfData(list)
}
\arguments{
\item{list}{A list of data frame with CNA data. Each data frame corresponds to a study.}
}
\value{
A data frame with node border attributes
}
\description{
Attribute CNA data to node border
}
\examples{
\dontrun{
load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
CNA_obj <- Node_obj_CNA_ProfData(ListCNAData_bkp)
}
}

