% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome.R
\name{Node_obj_CNA_ProfData}
\alias{Node_obj_CNA_ProfData}
\title{Attribute CNA data to node border}
\usage{
Node_obj_CNA_ProfData(list)
}
\arguments{
\item{list}{A list of data frame with CNA data. Each data frame corresponds to a study.}
}
\value{
A data frame with node border attributes
}
\description{
Attribute CNA data to node border
}
\examples{
\dontrun{
cgds <- CGDS("http://www.cbioportal.org/")
GeneList <- whichGeneList("DNA_damage_Response")
ProfDataCNA <- getProfileData.CGDS(cgds,GeneList, "brca_tcga_pub_gistic","brca_tcga_pub_all")
ListProfDataCNA <- list(ls1=ProfDataCNA, ls2=ProfDataCNA)
nodeObj <- Node_obj_CNA_ProfData(ListProfDataCNA)
}
}
