% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_doc.R
\name{Node_obj_FreqIn}
\alias{Node_obj_FreqIn}
\title{Attribute interaction frequency to node size}
\usage{
Node_obj_FreqIn(GeneList)
}
\arguments{
\item{GeneList}{A list of gene symbol}
}
\value{
A data frame with node attributes
}
\description{
Attribute interaction frequency to node size
}
\examples{
example <- "runManually"
\dontrun{
 load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
 node_obj <- Node_obj_FreqIn(GeneList)
}


}

