% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome.R
\name{Node_obj_FreqIn}
\alias{Node_obj_FreqIn}
\title{Attribute interaction frequency to node size}
\usage{
Node_obj_FreqIn(geneList)
}
\arguments{
\item{geneList}{A list of gene symbol}
}
\value{
A data frame with node attributes
}
\description{
Attribute interaction frequency to node size
}
\examples{
r_data <- new.env()
r_data[["FreqIn"]] <- structure(list(Genes = c("ATM", "ATR", "BRCA1", "BRCA2", "CHEK1",
"CHEK2", "FANCF", "MDC1", "RAD51"), FreqSum = c(0.04, 0.05, 0.05,
 0.03, 0.05, 0.04, 0.03, 0.03, 0.02)), .Names = c("Genes", "FreqSum"),
 class = "data.frame", row.names = c(NA, -9L))
 \dontrun{
GeneList <- whichGeneList("DNA_damage_Response")
nodeObj <- Node_obj_FreqIn(GeneList)
}

}
