% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_doc.R
\name{Node_obj_Met_ProfData}
\alias{Node_obj_Met_ProfData}
\title{Attribute gene Methylation to Nodes}
\usage{
Node_obj_Met_ProfData(list, type, Threshold)
}
\arguments{
\item{list}{a list of data frame with methylation data}

\item{type}{HM450 or HM27}

\item{Threshold}{the Rate cases (patients) that have a silencing genes by methylation}
}
\value{
a data frame with node shape attributes
}
\description{
Attribute gene Methylation to Nodes
}
\examples{

load(paste(path.package("bioCancer"),"/extdata/ListProfData.RData", sep=""))
Met_obj <- Node_obj_Met_ProfData(ListMetData_bkp$HM450,type="HM450",0.8)

}

