<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> bbe4d9cf03dcd14dc250f9c2de17d39c14fb1c1b
% Please edit documentation in R/Reactome_doc.R
\name{Node_obj_mRNA_Classifier}
\alias{Node_obj_mRNA_Classifier}
\title{Atrribute genes exprssion to color nodes}
\usage{
Node_obj_mRNA_Classifier(GeneList, GenesClassDetails = df)
}
\arguments{
\item{GeneList}{A gene list.}

\item{GenesClassDetails}{A dataframe with genes classes and genes expression.}
}
\value{
A data frame with node color attributes
}
\description{
Atrribute genes exprssion to color nodes
}
\examples{
\dontrun{
load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
Node_obj <- Node_obj_mRNA_Classifier(GeneList, GenesClassDetails)
}
}

