% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_doc.R
\name{Node_obj_mRNA_Classifier}
\alias{Node_obj_mRNA_Classifier}
\title{Atrribute genes expression to color nodes}
\usage{
Node_obj_mRNA_Classifier(GeneList,genesclassdetails)
}
\arguments{
\item{GeneList}{A gene list.}

\item{genesclassdetails}{A dataframe with genes classes and genes expression.}
}
\value{
A data frame with node color attributes
}
\description{
Atrribute genes expression to color nodes
}
\examples{
example <- "runManually"
\dontrun{
load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
node_obj <- Node_obj_mRNA_Classifier(GeneList, GenesClassDetails)
}

}

