% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffeewheel.R
\name{renderCoffeewheel}
\alias{renderCoffeewheel}
\title{Widget render function for use in Shiny}
\usage{
renderCoffeewheel(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

