% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich.R
\name{reStrColorGene}
\alias{reStrColorGene}
\title{Restructure the list of color attributed to the genes in every dimenssion for every studies}
\usage{
reStrColorGene(df)
}
\arguments{
\item{df}{data frame with colors attributed to the genes}
}
\value{
Hierarchical color attribute: gene > color
}
\description{
Restructure the list of color attributed to the genes in every dimenssion for every studies
}
\examples{
cgds <- cBioPortal(
hostname = "www.cbioportal.org",
protocol = "https",
api = "/api/v2/api-docs"
)
\dontrun{
getDataByGenes( api =  cgds,
studyId = "gbm_tcga_pub",
genes = c("NF1", "TP53", "ABL1"),
by = "hugoGeneSymbol",
molecularProfileIds = "gbm_tcga_pub_mrna"
)
}
}
