% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircomics_doc.R
\name{reStrColorGene}
\alias{reStrColorGene}
\title{Restructure the list of color attributed to the genes in every dimenssion for every studies}
\usage{
reStrColorGene(df)
}
\arguments{
\item{df}{data frame with colors attributed to the genes}
}
\value{
Hierarchical color attribute: gene > color
}
\description{
Restructure the list of color attributed to the genes in every dimenssion for every studies
}
\examples{
load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
Children <- reStrColorGene(ListProfData$Expression$luad_tcga_pub)

}

