% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich.R
\name{reStrDimension}
\alias{reStrDimension}
\title{Restructure the list of color attributed to the genes in every study for every dimensions}
\usage{
reStrDimension(LIST)
}
\arguments{
\item{LIST}{list of  hierarchical dimensions}
}
\value{
Hierarchical structure of: Study > dimensions > gene > color
}
\description{
Restructure the list of color attributed to the genes in every study for every dimensions
}
\examples{
cgds <- CGDS("http://www.cbioportal.org/")
\dontrun{
geneList <- whichGeneList("73")
ProfData <- getProfileData(cgds,
 geneList, "gbm_tcga_pub_mrna", "gbm_tcga_pub_all")
rownames(ProfData) <- NULL
TREE <- reStrDimension(list(
list1=list(df1=ProfData,df2=ProfData),
 list2=list(df3=ProfData,df4=ProfData)))
}
}
