% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCircomics_doc.R
\name{reStrDimension}
\alias{reStrDimension}
\title{Restructure the list of color attributed to the genes in every study for every dimensions}
\usage{
reStrDimension(LIST)
}
\arguments{
\item{LIST}{list of  hierarchical dimensions}
}
\value{
Hierarchical structure of: Study > dimensions > gene > color
}
\description{
Restructure the list of color attributed to the genes in every study for every dimensions
}
\examples{
\dontrun{
load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
Parent <- reStrDimension(ListProfData_bkp)
}
}

