% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircomics_doc.R
\name{reStrDimension}
\alias{reStrDimension}
\title{Restructure the list of color attributed to the genes in every study for every dimensions}
\usage{
reStrDimension(LIST)
}
\arguments{
\item{LIST}{list of  hierarchical dimensions}
}
\value{
Hierarchical structure of: Study > dimensions > gene > color
}
\description{
Restructure the list of color attributed to the genes in every study for every dimensions
}
\examples{
example <- "runManually"
\dontrun{
load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
Parent <- reStrDimension(ListProfData)
}

}

