<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> bbe4d9cf03dcd14dc250f9c2de17d39c14fb1c1b
% Please edit documentation in R/getCircomics_doc.R
\name{reStrDisease}
\alias{reStrDisease}
\title{Restructure the list of color attributed to the genes in every disease}
\usage{
reStrDisease(List)
}
\arguments{
\item{List}{of data frame with color attributes}
}
\value{
Hierarchy of dimensions in the same study: dimensions > gene > color
}
\description{
Restructure the list of color attributed to the genes in every disease
}
\examples{
\dontrun{
load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
Child <- reStrDisease(ListProfData_bkp$Expression)
}
}

