% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircomics_doc.R
\name{reStrDisease}
\alias{reStrDisease}
\title{Restructure the list of color attributed to the genes in every disease}
\usage{
reStrDisease(List)
}
\arguments{
\item{List}{of data frame with color attributes}
}
\value{
Hierarchy of dimensions in the same study: dimensions > gene > color
}
\description{
Restructure the list of color attributed to the genes in every disease
}
\examples{
example <- "runManually"
\dontrun{
load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
Child <- reStrDisease(ListProfData$Expression)
}


}

