% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich.R
\name{reStrDisease}
\alias{reStrDisease}
\title{Restructure the list of color attributed to the genes in every disease}
\usage{
reStrDisease(List)
}
\arguments{
\item{List}{of data frame with color attributes}
}
\value{
Hierarchy of dimensions in the same study: dimensions > gene > color
}
\description{
Restructure the list of color attributed to the genes in every disease
}
\examples{
\dontrun{
cgds <- CGDS("http://www.cbioportal.org/")
geneList <- whichGeneList("73")
ProfData <- getProfileData.CGDS(cgds,
 geneList, "gbm_tcga_pub_mrna", "gbm_tcga_pub_all")
rownames(ProfData) <- NULL
tree <- reStrDisease(list(df1=ProfData,df2=ProfData))
}
}
