% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_doc.R
\name{Studies_obj}
\alias{Studies_obj}
\title{get object for grViz. Link Studies to genes}
\usage{
Studies_obj(df)
}
\arguments{
\item{df}{data frame with gene classes}
}
\value{
grViz object. a data frame with Study attributes
}
\description{
get object for grViz. Link Studies to genes
}
\examples{
example <- "runManually"
\dontrun{
 string1 <- "https://wiki.ubuntu.com/kmezhoud/bioCancer?"
 string2 <- "action=AttachFile&do=get&target=ListProfData.RData"
 link <- curl::curl(paste0(string1,string2, sep=""))
 load(link)
##load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
ReactomeFI <- readRDS(paste0(system.file(package = "bioCancer"),"/extdata/DisGeNet.RDS"))
Sd_obj <- Studies_obj(GenesClassDetails)
}
}

