% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome.R
\name{Studies_obj}
\alias{Studies_obj}
\title{get object for grViz. Link Studies to genes}
\usage{
Studies_obj(df)
}
\arguments{
\item{df}{data frame with gene classes}
}
\value{
grViz object. a data frame with Study attributes
}
\description{
get object for grViz. Link Studies to genes
}
\examples{
Studies_obj(data.frame("col1", "col2", "col3", "col4", "col5", "col6"))
\dontrun{
Genes ranking     class postProb exprsMeanDiff exprsUpDw
1 FANCF       1 brca_tcga  1.00000      179.9226        UP
2  MLH1       1  gbm_tcga  0.99703      256.3173        UP
}
}
