<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> bbe4d9cf03dcd14dc250f9c2de17d39c14fb1c1b
% Please edit documentation in R/Reactome_doc.R
\name{Studies_obj}
\alias{Studies_obj}
\title{get object for grViz. Link Studies to genes}
\usage{
Studies_obj(df = df)
}
\arguments{
\item{df}{data frame with gene classes}
}
\value{
grViz object. a data frame with Study attributes
}
\description{
get object for grViz. Link Studies to genes
}
\examples{
\dontrun{
load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
Sd_obj <- Studies_obj(GenesClassDetails)
}
}

