% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircomics_doc.R
\name{UnifyRowNames}
\alias{UnifyRowNames}
\title{Unify row names in data frame with the same order of gene list.}
\usage{
UnifyRowNames(x, GeneList)
}
\arguments{
\item{x}{data frame with gene symbol in the row name}

\item{GeneList}{a gene list}
}
\value{
a data frame having the gene in row name ordered as in gene list.
}
\description{
Unify row names in data frame with the same order of gene list.
}
\examples{
load(paste(path.package("bioCancer"),"/extdata/ListProfData.RData", sep=""))
Freq_Mut <- UnifyRowNames(ListMutData_bkp$luad_tcga_pub, GeneList = GeneList)

}

