% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich.R
\name{UnifyRowNames}
\alias{UnifyRowNames}
\title{Unify row names in data frame with the same order of gene list.}
\usage{
UnifyRowNames(x,geneList)
}
\arguments{
\item{x}{data frame with gene symbol in the row name}

\item{geneList}{a gene list}
}
\value{
a data frame having the gene in row name ordered as in gene list.
}
\description{
Unify row names in data frame with the same order of gene list.
}
\examples{
\dontrun{
cgds <- CGDS("http://www.cbioportal.org/")
geneList <- whichGeneList("73")
ProfData <- getProfileData.CGDS(cgds,
 geneList, "gbm_tcga_pub_mrna", "gbm_tcga_pub_all")
rownames(ProfData) <- NULL
geneListOrder <- UnifyRowNames(list(
list1=list(df1=ProfData,df2=ProfData),
 list2=list(df3=ProfData,df4=ProfData)),
  geneList)
  }
}
