% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircomics_doc.R
\name{whichGeneList}
\alias{whichGeneList}
\title{Identify which gene list in input}
\usage{
whichGeneList()
}
\value{
gene list
}
\description{
Identify which gene list in input
}
\examples{
example <- "runManually"
\dontrun{
load(paste(system.file(package="bioCancer"),"/extdata/ListProfData.RData", sep=""))
GeneList <- whichGeneList()
}

}

