<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> bbe4d9cf03dcd14dc250f9c2de17d39c14fb1c1b
% Please edit documentation in R/getCircomics_doc.R
\name{whichGeneList}
\alias{whichGeneList}
\title{Identify which gene list in input}
\usage{
whichGeneList()
}
\value{
gene list
}
\description{
Identify which gene list in input
}
\examples{
\dontrun{
load(paste(path.package("bioCancer"),"/data/ListProfData.RData", sep=""))
GeneList <- whichGeneList()
}
}

